+++ arcade/examples/gui/0_basic_setup.py
@@ -1,6 +1,6 @@
 """Demonstrates general setup.
 
-If arcade and Python are properly installed, you can run this example with:
+If Arcade and Python are properly installed, you can run this example with:
 python -m arcade.examples.gui.0_basic_setup
 
 Content: