+++ arcade/examples/gui/2_widgets.py
@@ -2,7 +2,7 @@
 
 See the other GUI examples for more indepth information about specific widgets.
 
-If arcade and Python are properly installed, you can run this example with:
+If Arcade and Python are properly installed, you can run this example with:
 python -m arcade.examples.gui.2_widgets
 """