+++ arcade/examples/sprite_follow_simple_2.py
@@ -31,7 +31,7 @@ SPRITE_SPEED = 0.5
 class Coin(arcade.Sprite):
     """
     This class represents the coins on our screen. It is a child class of
-    the arcade library's "Sprite" class.
+    the Arcade library's "Sprite" class.
     """
 
     def follow_sprite(self, player_sprite):